Added some buffering to the socket
The socket now buffers notes for 200ms when multiple notes are toggled
by dragging. This is to try and reduce latency when running on the
internet by reducing the number of messages between the client and
server. Also moved the code for the timeout for the volume into the
socket file.