Improved Setup of multigrid functionality
Fixed the way the grids are presented so that the leftmost grid always
belongs to the local user. This involved redoing the whole setup of the
grid in script.js, so grids can be added dynamically. Added a page for
the user to specify their name. Technically the number of grids is now
unlimited, but more than 2 hasnt been sorted out properly yet.